requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic-settings
python-dotenv
langchain-google-genai
langgraph
google-api-python-client
google-auth-oauthlib
pymongo
motor
pytz
passlib[bcrypt]
python-jose[cryptography]
requests
python-json-logger
pydantic[email]
python-multipart